# Ensure `dtype=object` arrays always have `int` objects, even after assignment