--- a/tests/fields/test_assignment.py
+++ b/tests/fields/test_assignment.py
@@ -22,6 +22,14 @@ class TestConstantIndex:
         with pytest.raises(ValueError):
             a[0] = field.order
 
+    def test_always_int_object(self):
+        # Ensure when assigning FieldArray elements to an array they are converted to ints
+        GF = galois.GF(2**100)
+        a = GF.Random(10)
+        assert np.all(is_int(a))
+        a[0] = GF(10)
+        assert np.all(is_int(a))
+
 
 class TestSliceIndex:
     def test_constant_valid(self, field):
@@ -70,6 +78,14 @@ class TestSliceIndex:
         with pytest.raises(ValueError):
             a[0:2] = np.array([field.order, 1])
 
+    def test_always_int_object(self):
+        # Ensure when assigning FieldArray elements to an array they are converted to ints
+        GF = galois.GF(2**100)
+        a = GF.Random(10)
+        assert np.all(is_int(a))
+        a[0:3] = [GF(10), GF(20), GF(30)]
+        assert np.all(is_int(a))
+
 
 class Test2DSliceIndex:
     def test_list_valid(self, field):
@@ -103,3 +119,14 @@ class Test2DSliceIndex:
         a = field.Random((10,10))
         with pytest.raises(ValueError):
             a[0:2, 0:2] = np.array([[field.order, 1], [1, 1]])
+
+    def test_always_int_object(self):
+        # Ensure when assigning FieldArray elements to an array they are converted to ints
+        GF = galois.GF(2**100)
+        a = GF.Random((10,10))
+        assert np.all(is_int(a))
+        a[0:2, 0:2] = [[GF(10), GF(20)], [GF(30), GF(40)]]
+        assert np.all(is_int(a))
+
+
+is_int = np.vectorize(lambda element: isinstance(element, int))