This PR cleans up element assignment and verification and removes the need to cast every ufunc input to an integer. Fixes #321

There is a small performance benefit, too.

```python
In [1]: import galois

In [2]: GF = galois.GF(109987**4)

In [3]: x = GF.Random((1000), seed=1)

In [4]: y = GF.Random((1000), seed=2)

In [5]: x * y;

# Before
In [7]: %timeit x * y;
56.2 ms ± 174 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# After
In [6]: %timeit x * y;
55.3 ms ± 112 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```